Schedules probe engine execution
