+++ src/main/java/io/jenkins/pluginhealth/scoring/schedule/UpdateCenterScheduler.java
@@ -1,0 +1,60 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2022 Jenkins Infra
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.jenkins.pluginhealth.scoring.schedule;
+
+import java.io.IOException;
+
+import io.jenkins.pluginhealth.scoring.probes.ProbeEngine;
+import io.jenkins.pluginhealth.scoring.service.PluginService;
+import io.jenkins.pluginhealth.scoring.service.UpdateCenterService;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+@Component
+public class UpdateCenterScheduler implements CommandLineRunner {
+    private final UpdateCenterService updateCenterService;
+    private final PluginService pluginService;
+    private final ProbeEngine probeEngine;
+
+    public UpdateCenterScheduler(UpdateCenterService updateCenterService, PluginService pluginService,
+                                 ProbeEngine probeEngine) {
+        this.updateCenterService = updateCenterService;
+        this.pluginService = pluginService;
+        this.probeEngine = probeEngine;
+    }
+
+    @Override
+    public void run(String... args) throws IOException {
+        updateDatabase();
+    }
+
+    @Scheduled(cron = "${cron.update-center}", zone = "UTC")
+    public void updateDatabase() throws IOException {
+        updateCenterService.readUpdateCenter().forEach(pluginService::saveOrUpdate);
+        probeEngine.run();
+    }
+}
